Code refactoring.

+++ src/p_box.rs
@@ -1,16 +1,15 @@
 use std::vec;
 
-
-pub fn generate_pbox(key: String, size: usize) -> Result<Vec<u64>, String> {
+pub fn generate_pbox(key: String, size: usize) -> Result<Vec<u8>, String> {
     if key.is_empty() {
         return Err(String::from("Key is empty"));
     }
     if size <= 0 {
         return Err(String::from("Size is lower/equal zero"));
     }
-    let mut result: Vec<u64> = vec![];
+    let mut result: Vec<u8> = vec![];
     for i in 0..size {
-        result.push(i as u64);
+        result.push(i as u8);
     }
     let key_len = key.len();
     for i in 0..result.len() {
@@ -20,23 +19,23 @@ pub fn generate_pbox(key: String, size: usize) -> Result<Vec<u64>, String> {
     Ok(result)
 }
 
-pub fn encrypt_pbox(input_str: String, p_box: Vec<u64>) -> Result<Vec<u64>, String> {
+pub fn encrypt_pbox(input_str: String, p_box: Vec<u8>) -> Result<Vec<u8>, String> {
     if input_str.is_empty() {
         return Err(String::from("Input string is empty"));
     }
     if p_box.is_empty() {
         return Err(String::from("P-box is empty"));
     }
-    let mut result: Vec<u64> = vec![];
+    let mut result: Vec<u8> = vec![];
     let str_bytes = input_str.clone().into_bytes();
     for i in 0..p_box.len() {
-        let temp = str_bytes[p_box[i] as usize] as u64;
+        let temp = str_bytes[p_box[i] as usize] as u8;
         result.push(temp);
     }
     Ok(result)
 }
 
-pub fn decrypt_pbox(input: Vec<u64>, p_box: Vec<u64>) -> Result<String, String> {
+pub fn decrypt_pbox(input: Vec<u8>, p_box: Vec<u8>) -> Result<String, String> {
     if input.is_empty() {
         return Err(String::from("Input is empty"));
     }
@@ -53,17 +52,17 @@ pub fn decrypt_pbox(input: Vec<u64>, p_box: Vec<u64>) -> Result<String, String> 
     Ok(result)
 }
 
-pub fn invers_pbox(input: Vec<u64>) -> Result<Vec<u64>, String> {
+pub fn invers_pbox(input: Vec<u8>) -> Result<Vec<u8>, String> {
     if input.is_empty() {
         return Err(String::from("P-box is empty"));
     }
-    let mut result: Vec<u64> = vec![];
+    let mut result: Vec<u8> = vec![];
     let size = input.len();
     for i in 0..size {
-        result.push(i as u64);
+        result.push(i as u8);
     }
     for (index, value) in input.iter().enumerate() {
-        result[*value as usize] = index as u64;
+        result[*value as usize] = index as u8;
     }
     Ok(result)
 }
